fix failing test. move curved text to text module.
